don't use conduit for http requests.
it somehow resulted in either leaks or just weird memory usage patterns.

there is still a problem, though, because `fetchAll` uses ~30mb
throughout the fetching phase, but then suddenly jumps to >500mb when
writing to the file. this happens with both http-conduit and standard
http, so it might be an ordinary lazyness problem, but i'm not sure why
it happens. need to reduce area i have to test for this.